Project commit

+++ database/migrations/2020_01_28_195156_add_language_options.php
@@ -14,10 +14,18 @@ class AddLanguageOptions extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('card_brand');
-            $table->dropColumn('stripe_id');
-            $table->dropColumn('card_last_four');
-            $table->dropColumn('trial_ends_at');
+            if (Schema::hasColumn('users', 'card_brand')) {
+                $table->dropColumn('card_brand');   
+            }
+            if (Schema::hasColumn('users', 'stripe_id')) {
+                $table->dropColumn('stripe_id');   
+            }
+            if (Schema::hasColumn('users', 'card_last_four')) {
+                $table->dropColumn('card_last_four');
+            }
+            if (Schema::hasColumn('users', 'trial_ends_at')) {
+                $table->dropColumn('trial_ends_at');
+            }
             $table->string("language", 2)->default("EN")->after("remember_token");
         });
